requirements.txt
streamlit
watchdog
langchain
python-dotenv
langchain-google-vertexai
